--- a/.gitignore
+++ b/.gitignore
@@ -5,3 +5,4 @@ fp-info-cache
 .vscode
 .claude
 *-bak
+litter_box.md