--- a/firmware/include/dma_automatic_validation.h
+++ b/firmware/include/dma_automatic_validation.h
@@ -63,9 +63,9 @@ public:
         uint32_t trigger_duration_us;  // Trigger pulse duration
         
         external_trigger_config_t() :
-            trigger_pin(33), enable_logic_analyzer(false), 
-            enable_oscilloscope(false), trigger_delay_us(0),
-            trigger_duration_us(10) {}
+            trigger_pin( 33 ), enable_logic_analyzer( false ), 
+            enable_oscilloscope( false ), trigger_delay_us( 0 ),
+            trigger_duration_us( 10 ) {}
     };
     
     // Environmental monitoring configuration
@@ -81,12 +81,12 @@ public:
         uint8_t humidity_max_percent;
         
         environmental_config_t() :
-            enable_temperature_monitoring(true),
-            enable_voltage_monitoring(true),
-            enable_humidity_monitoring(false),
-            temperature_min_c(-20.0f), temperature_max_c(70.0f),
-            voltage_min_v(4.5f), voltage_max_v(5.5f),
-            humidity_min_percent(10), humidity_max_percent(90) {}
+            enable_temperature_monitoring( true ),
+            enable_voltage_monitoring( true ),
+            enable_humidity_monitoring( false ),
+            temperature_min_c( -20.0f ), temperature_max_c( 70.0f ),
+            voltage_min_v( 4.5f ), voltage_max_v( 5.5f ),
+            humidity_min_percent( 10 ), humidity_max_percent( 90 ) {}
     };
     
     // Comprehensive test configuration
@@ -104,15 +104,15 @@ public:
         environmental_config_t env_config;
         
         validation_config_t() :
-            current_phase(validation_phase_t::PHASE_1_IMMEDIATE),
-            enable_automatic_progression(true),
-            enable_data_logging(true),
-            enable_serial_reporting(true),
-            enable_led_diagnostics(true),
-            enable_external_triggers(false),
-            enable_environmental_monitoring(true),
-            phase_duration_ms(86400000),  // 24 hours default
-            test_interval_ms(60000) {}    // 1 minute between tests
+            current_phase( validation_phase_t::PHASE_1_IMMEDIATE ),
+            enable_automatic_progression( true ),
+            enable_data_logging( true ),
+            enable_serial_reporting( true ),
+            enable_led_diagnostics( true ),
+            enable_external_triggers( false ),
+            enable_environmental_monitoring( true ),
+            phase_duration_ms( 86400000 ),  // 24 hours default
+            test_interval_ms( 60000 ) {}    // 1 minute between tests
     };
     
     // Validation results structure
@@ -137,12 +137,12 @@ public:
         uint8_t humidity_percent;
         
         validation_result_t() :
-            phase(validation_phase_t::PHASE_1_IMMEDIATE),
-            test_id(0), test_name(nullptr), passed(false),
-            timestamp_ms(0), duration_ms(0), operations_completed(0),
-            errors_encountered(0), error_rate_ppm(0.0f),
-            efficiency_percentage(0.0f), failure_reason(nullptr),
-            temperature_c(25.0f), voltage_v(5.0f), humidity_percent(50) {}
+            phase( validation_phase_t::PHASE_1_IMMEDIATE ),
+            test_id( 0 ), test_name( nullptr ), passed( false ),
+            timestamp_ms( 0 ), duration_ms( 0 ), operations_completed( 0 ),
+            errors_encountered( 0 ), error_rate_ppm( 0.0f ),
+            efficiency_percentage( 0.0f ), failure_reason( nullptr ),
+            temperature_c( 25.0f ), voltage_v( 5.0f ), humidity_percent( 50 ) {}
     };
     
     // Acceptance criteria structure (from knowledge base)
@@ -219,7 +219,7 @@ private:
     
     // Threading
     Threads::Mutex validation_mutex_;
-    Threads::ThreadWait validation_thread_wait_;
+    volatile bool should_exit_thread_;
     
     // Data logging
     File log_file_;
@@ -246,53 +246,53 @@ private:
         float overall_error_rate;
         
         validation_statistics_t() :
-            total_tests_run(0), total_tests_passed(0),
-            total_operations(0), total_errors(0),
-            total_duration_ms(0), overall_efficiency(0.0f),
-            overall_error_rate(0.0f) {}
+            total_tests_run( 0 ), total_tests_passed( 0 ),
+            total_operations( 0 ), total_errors( 0 ),
+            total_duration_ms( 0 ), overall_efficiency( 0.0f ),
+            overall_error_rate( 0.0f ) {}
     } statistics_;
     
     // Internal methods
     void validation_thread_function();
-    static void validation_thread_entry(void* user_data);
+    static void validation_thread_entry( void* user_data );
     
     // Phase 1 test implementations
-    validation_result_t run_phase1_basic_functionality();
-    validation_result_t run_phase1_stress_testing();
-    validation_result_t run_phase1_concurrent_operations();
-    validation_result_t run_phase1_performance_baseline();
-    validation_result_t run_phase1_error_injection();
-    validation_result_t run_phase1_long_term_stability();
+    dma_automatic_validation::validation_result_t run_phase1_basic_functionality();
+    dma_automatic_validation::validation_result_t run_phase1_stress_testing();
+    dma_automatic_validation::validation_result_t run_phase1_concurrent_operations();
+    dma_automatic_validation::validation_result_t run_phase1_performance_baseline();
+    dma_automatic_validation::validation_result_t run_phase1_error_injection();
+    dma_automatic_validation::validation_result_t run_phase1_long_term_stability();
     
     // Phase 2 test implementations (external equipment)
-    validation_result_t run_phase2_timing_analysis();
-    validation_result_t run_phase2_performance_measurement();
-    validation_result_t run_phase2_signal_integrity();
-    validation_result_t run_phase2_environmental_stress();
-    validation_result_t run_phase2_emc_compliance();
-    validation_result_t run_phase2_power_efficiency();
+    dma_automatic_validation::validation_result_t run_phase2_timing_analysis();
+    dma_automatic_validation::validation_result_t run_phase2_performance_measurement();
+    dma_automatic_validation::validation_result_t run_phase2_signal_integrity();
+    dma_automatic_validation::validation_result_t run_phase2_environmental_stress();
+    dma_automatic_validation::validation_result_t run_phase2_emc_compliance();
+    dma_automatic_validation::validation_result_t run_phase2_power_efficiency();
     
     // Phase 3 test implementations
-    validation_result_t run_phase3_production_testing();
-    validation_result_t run_phase3_field_diagnostics();
-    validation_result_t run_phase3_calibration_procedures();
-    validation_result_t run_phase3_quality_control();
+    dma_automatic_validation::validation_result_t run_phase3_production_testing();
+    dma_automatic_validation::validation_result_t run_phase3_field_diagnostics();
+    dma_automatic_validation::validation_result_t run_phase3_calibration_procedures();
+    dma_automatic_validation::validation_result_t run_phase3_quality_control();
     
     // Utility methods
-    bool check_acceptance_criteria(const validation_result_t& result);
-    void log_result_to_sd(const validation_result_t& result);
-    void print_result_to_serial(const validation_result_t& result);
-    void update_led_status(validation_phase_t phase, bool test_active, bool test_passed);
+    bool check_acceptance_criteria( const validation_result_t& result );
+    void log_result_to_sd( const validation_result_t& result );
+    void print_result_to_serial( const validation_result_t& result );
+    void update_led_status( validation_phase_t phase, bool test_active, bool test_passed );
     void trigger_external_equipment();
     void read_environmental_sensors();
-    void update_statistics(const validation_result_t& result);
+    void update_statistics( const validation_result_t& result );
     
 public:
-    dma_automatic_validation(dma_performance_validator* validator,
+    dma_automatic_validation( dma_performance_validator* validator,
                             dma_test_suite* test_suite,
                             dma_realtime_monitor* monitor,
                             dma_diagnostics::dma_error_handler* error_handler,
-                            const validation_config_t& config = validation_config_t());
+                            const validation_config_t& config = validation_config_t() );
     ~dma_automatic_validation();
     
     // Validation control
@@ -301,39 +301,39 @@ public:
     bool is_validation_active() const { return validation_active_; }
     
     // Phase control
-    void set_validation_phase(validation_phase_t phase);
+    void set_validation_phase( validation_phase_t phase );
     validation_phase_t get_current_phase() const { return current_phase_; }
     bool advance_to_next_phase();
     
     // Test execution
-    validation_result_t run_single_test(validation_phase_t phase, uint8_t test_id);
+    validation_result_t run_single_test( validation_phase_t phase, uint8_t test_id );
     bool run_current_phase_tests();
     bool run_all_phases();
     
     // Results access
-    const validation_result_t* get_results(uint16_t& count) const;
+    const validation_result_t* get_results( uint16_t& count ) const;
     const validation_result_t& get_latest_result() const;
     const validation_statistics_t& get_statistics() const { return statistics_; }
     
     // Acceptance criteria
-    bool check_phase_acceptance(validation_phase_t phase) const;
+    bool check_phase_acceptance( validation_phase_t phase ) const;
     bool check_overall_acceptance() const;
     const acceptance_criteria_t& get_acceptance_criteria() const { return criteria_; }
-    void set_acceptance_criteria(const acceptance_criteria_t& criteria) { criteria_ = criteria; }
+    void set_acceptance_criteria( const acceptance_criteria_t& criteria ) { criteria_ = criteria; }
     
     // Configuration
-    void set_config(const validation_config_t& config) { config_ = config; }
+    void set_config( const validation_config_t& config ) { config_ = config; }
     const validation_config_t& get_config() const { return config_; }
     
     // External equipment integration
-    void enable_external_triggers(const external_trigger_config_t& trigger_config);
+    void enable_external_triggers( const external_trigger_config_t& trigger_config );
     void disable_external_triggers();
     void manual_trigger();
     
     // Environmental monitoring
-    void enable_environmental_monitoring(const environmental_config_t& env_config);
+    void enable_environmental_monitoring( const environmental_config_t& env_config );
     void disable_environmental_monitoring();
-    void get_environmental_conditions(float& temperature, float& voltage, uint8_t& humidity) const;
+    void get_environmental_conditions( float& temperature, float& voltage, uint8_t& humidity ) const;
     
     // Data logging
     bool initialize_sd_logging();
@@ -342,13 +342,13 @@ public:
     
     // Reporting
     void print_validation_summary() const;
-    void print_phase_report(validation_phase_t phase) const;
+    void print_phase_report( validation_phase_t phase ) const;
     void print_detailed_statistics() const;
     void export_results_to_csv();
     
     // Utility methods
-    static const char* phase_to_string(validation_phase_t phase);
-    static const char* get_test_name(validation_phase_t phase, uint8_t test_id);
+    static const char* phase_to_string( validation_phase_t phase );
+    static const char* get_test_name( validation_phase_t phase, uint8_t test_id );
     uint32_t get_estimated_completion_time() const;
     float get_progress_percentage() const;
 };