… claude/contexts/data_store.json
… claude/data_store/knowledge/firmware/dma_automated_validation_code_modifications.json
… claude/data_store/knowledge/firmware/dma_automated_validation_plan.json
… claude/data_store/knowledge/firmware/dma_automatic_validation_implementation.json
… claude/data_store/knowledge/firmware/dma_hardware_testing_plan.json
… claude/data_store/knowledge/firmware/dma_hardware_testing_validation.json
… claude/data_store/knowledge/firmware/dma_i2c_architecture_design.json
… claude/data_store/knowledge/firmware/dma_i2c_complete_implementation.json
… claude/data_store/knowledge/firmware/dma_phase1_execution_plan.json
… claude/data_store/knowledge/firmware/dma_testing_plan_knowledge.json
… claude/data_store/knowledge/firmware/imxrt1062_i2c_dma.json
… documentation/README.md
… documentation/firmware/DMA_VALIDATION_SYSTEM.md
… documentation/firmware/FIRMWARE_ARCHITECTURE.md
… firmware/include/dma_error_handler.h
… firmware/include/dma_performance_validator.h
… firmware/include/drivers/dma_i2c_hal.h
… firmware/include/drivers/rob_tillaart_ad_5993r_async.h
… firmware/include/electric_mayhem_dma.h
… firmware/src/dma_automatic_validation.cpp
… firmware/src/dma_error_handler.cpp
… firmware/src/dma_performance_validator.cpp
… firmware/src/drivers/dma_i2c_hal.cpp
… firmware/src/drivers/rob_tillaart_ad_5993r_async.cpp
… firmware/src/main.cpp